Reorganize extension code into src directory structure
- Move main extension files from root to organized src/ structure
- Create logical subdirectories: src/core/, src/tasks/, src/fuzzing/
- Update package.json main entry point to src/extension.js
- Update all import/require statements in moved files
- Update all test file imports to reference new file locations
- Maintain full functionality and test coverage (49 tests passing)

Files moved:
- extension.js → src/extension.js
- dockerOperations.js → src/core/dockerOperations.js
- taskProvider.js → src/tasks/taskProvider.js

This creates a more professional and maintainable codebase structure
while preserving all existing functionality.
